Switch the aggregate component base to a template class.
