+++ src/compy/index.js
@@ -1,7 +1,6 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '#x/BaseAggregateComponent';
 export * from '#x/BaseComponent';
 export * from '#x/BaseConfig';
 export * from '#x/BaseRootComponent';
@@ -9,4 +8,5 @@ export * from '#x/BaseWrappedHierarchy';
 export * from '#x/ControlContext';
 export * from '#x/Names';
 export * from '#x/RootControlContext';
+export * from '#x/TemplAggregateComponent';
 export * from '#x/TemplThreadComponent';
