+++ src/webapp-builtins/tests/NopComponent.js
@@ -1,14 +1,14 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseAggregateComponent } from '@this/compy';
+import { BaseComponent, TemplAggregateComponent } from '@this/compy';
 
 
 /**
  * Minimal concrete component class, which has no-op implementations for all
  * `_impl_*` methods.
  */
-export class NopComponent extends BaseAggregateComponent {
+export class NopComponent extends TemplAggregateComponent('NopAggregate', BaseComponent) {
   // @defaultConstructor
 
   /** @override */
